useReducer

+++ src/contexts/CitiesContext.tsx
@@ -1,6 +1,73 @@
-import { createContext, useEffect, useState, ReactNode, useContext } from "react";
+// ***--- useReducer ---***
+
+import { createContext, useEffect, ReactNode, useContext, useReducer } from "react";
 
 const BASE_URL = "http://localhost:9000";
+
+interface StateType {
+    cities: CityContextType[];
+    isLoading: boolean;
+    currentCity: CityContextType | null;
+    error: string;
+}
+
+type ActionType =
+    | { type: "loading" }
+    | { type: "cities/loaded"; payload: CityContextType[] }
+    | { type: "city/loaded"; payload: CityContextType }
+    | { type: "city/created"; payload: CityContextType }
+    | { type: "city/deleted"; payload: string }
+    | { type: "rejected"; payload: string };
+
+function reducer(state: StateType, action: ActionType): StateType {
+    switch (action.type) {
+        case "loading":
+            return { ...state, isLoading: true };
+
+        case "cities/loaded":
+            return {
+                ...state,
+                isLoading: false,
+                cities: action.payload,
+            };
+
+        case "city/loaded":
+            return {
+                ...state,
+                isLoading: false,
+                currentCity: action.payload,
+            };
+
+        case "city/created":
+            return {
+                ...state,
+                isLoading: false,
+                cities: [...state.cities, action.payload],
+                currentCity: action.payload,
+            };
+
+        case "city/deleted":
+            return {
+                ...state,
+                isLoading: false,
+                cities: state.cities.filter((city) => city.id !== action.payload),
+                currentCity: null,
+            };
+
+        case "rejected":
+            return { ...state, isLoading: false, error: action.payload };
+
+        default:
+            throw new Error("Unknown action type");
+    }
+}
+
+const initialState = {
+    cities: [],
+    isLoading: false,
+    currentCity: null,
+    error: "",
+};
 
 export interface CityContextType {
     cityName: string;
@@ -19,6 +86,7 @@ export interface CitiesContextType {
     cities: CityContextType[];
     isLoading: boolean;
     currentCity: CityContextType | null;
+    error: string;
     getCity: (id: string | undefined) => Promise<void>;
     createCity: (newCity: CityContextType) => Promise<void>;
     deleteCity: (id: string | undefined) => Promise<void>;
@@ -27,21 +95,22 @@ export interface CitiesContextType {
 const CitiesContext = createContext<CitiesContextType | undefined>(undefined);
 
 export const CitiesProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
-    const [cities, setCities] = useState<CityContextType[]>([]);
-    const [isLoading, setIsLoading] = useState(false);
-    const [currentCity, setCurrentCity] = useState(null);
+    // const [state, dispatch] = useReducer(reducer, initialState);
+    const [{ cities, isLoading, currentCity, error }, dispatch]: [StateType, React.Dispatch<ActionType>] = useReducer(
+        reducer,
+        initialState
+    );
 
     useEffect(() => {
         async function fetchCities() {
+            dispatch({ type: "loading" });
+
             try {
-                setIsLoading(true);
                 const res = await fetch(`${BASE_URL}/cities`);
                 const data = await res.json();
-                setCities(data);
+                dispatch({ type: "cities/loaded", payload: data });
             } catch (error) {
-                alert("There was an error loading data...");
-            } finally {
-                setIsLoading(false);
+                dispatch({ type: "rejected", payload: "There was an error loading cities..." });
             }
         }
 
@@ -49,23 +118,24 @@ export const CitiesProvider: React.FC<{ children: ReactNode }> = ({ children }) 
     }, []);
 
     async function getCity(id: string | undefined) {
+        if (currentCity && id === currentCity.id) {
+            return;
+        }
+        dispatch({ type: "loading" });
+
         try {
-            setIsLoading(true);
-            setCurrentCity(null);
             const res = await fetch(`${BASE_URL}/cities/${id}`);
             const data = await res.json();
-            setCurrentCity(data);
+            dispatch({ type: "city/loaded", payload: data });
         } catch (error) {
-            alert("There was an error loading data...");
-        } finally {
-            setIsLoading(false);
+            dispatch({ type: "rejected", payload: "There was an error loading city..." });
         }
     }
 
     async function createCity(newCity: CityContextType) {
+        dispatch({ type: "loading" });
+
         try {
-            setIsLoading(true);
-            setCurrentCity(null);
             const res = await fetch(`${BASE_URL}/cities`, {
                 method: "POST",
                 body: JSON.stringify(newCity),
@@ -74,32 +144,32 @@ export const CitiesProvider: React.FC<{ children: ReactNode }> = ({ children }) 
                 },
             });
             const data = await res.json();
-            setCities((cities) => [...cities, data]);
+            dispatch({ type: "city/created", payload: data });
         } catch (error) {
-            alert("There was an error creating city...");
-        } finally {
-            setIsLoading(false);
+            dispatch({ type: "rejected", payload: "There was an error creating city..." });
         }
     }
 
     async function deleteCity(id: string | undefined) {
-        try {
-            setIsLoading(true);
-            setCurrentCity(null);
-            await fetch(`${BASE_URL}/cities/${id}`, {
-                method: "DELETE",
-            });
+        dispatch({ type: "loading" });
 
-            setCities((cities) => cities.filter((city) => city.id !== id));
+        try {
+            if (id !== undefined) {
+                // Check if id is not undefined
+                await fetch(`${BASE_URL}/cities/${id}`, {
+                    method: "DELETE",
+                });
+                dispatch({ type: "city/deleted", payload: id });
+            } else {
+                throw new Error("Cannot delete city: ID is undefined");
+            }
         } catch (error) {
-            alert("There was an error deleting city...");
-        } finally {
-            setIsLoading(false);
+            dispatch({ type: "rejected", payload: "There was an error deleting city..." });
         }
     }
 
     return (
-        <CitiesContext.Provider value={{ cities, isLoading, currentCity, getCity, createCity, deleteCity }}>
+        <CitiesContext.Provider value={{ cities, isLoading, currentCity, error, getCity, createCity, deleteCity }}>
             {children}
         </CitiesContext.Provider>
     );
